+++ src/lidia/config.py
@@ -1,4 +1,12 @@
+import os
+from os import path
+from typing import Optional
+from pydantic import Field
+
 from .mytypes import NestingModel
+
+
+g_config: 'Config' = None
 
 
 class RpctaskConfig(NestingModel):
@@ -24,5 +32,19 @@ class Config(NestingModel):
     Every config field in this and children should be provided with defaults,
     that can be overriden by config files"""
 
+    dollar_schema: Optional[str] = Field(alias='$schema', default=None)
+    """Allow the `$schema` property for specifying JSON Schema URL"""
     rpctask = RpctaskConfig()
     approach = ApproachConfig()
+
+
+def schema_location():
+    root_path = path.abspath(path.dirname(__file__))
+    data_path = path.join(root_path, 'data')
+    os.makedirs(data_path, exist_ok=True)
+    return path.join(data_path, 'lidia-config.json')
+
+
+def write_schema():
+    with open(schema_location(), 'w') as out:
+        out.write(Config.schema_json(by_alias=True))